move messages to containers
